Flatten module hierarchy
