+++ src/pso.rs
@@ -9,7 +9,7 @@ pub mod builder;
 use crate::pso::probe::console_probe::ConsoleProbe;
 use crate::pso::probe::Probe;
 use crate::pso::swarm::Swarm;
-use crate::test_functions::test_functions::rosenbrock;
+use crate::test_functions::rosenbrock;
 
 
 /// Struct wrapping all configuration parameters of PSO algorithm.
